requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
prometheus-client
gunicorn
gevent
